define expectedNodeCount inside subject node; move functions back into helpers.go
